Fixed a bunch of warnings, disabled some of them on GCC.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7479 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/Common.cpp
@@ -53,7 +53,7 @@ public:
 			// BluePill #4539 - show whether this is a 32-bit or 64-bit binary
 			sprintf( string, "%s/%u #%d (%s)%s %s %s %s",
 				ENGINE_VERSION,
-				sizeof(void*) * 8,
+				(int)sizeof(void*) * 8,
 				RevisionTracker::Instance().GetHighestRevision(), RevisionTracker::Instance().GetRevisionString(),
 				BUILD_DEBUG, BUILD_STRING, __DATE__, __TIME__
 			);
